requirements.txt
simplejson==3.17.3
Flask==2.0.1
mysql-connector-python==8.0.26
pandas==1.3.1